requirements.txt
selenium
webdriver-manager
pandas
beautifulsoup4
openpyxl
python-dotenv